Rename src/search/ to src/web-search/
