--- src/search/index.tsx
@@ -1,2 +1,0 @@
-// Main entry point
-export { default as LiveSearchInput } from './containers/LiveSearchInput'
