+++ src/web-search/LiveSearchInput.module.css
@@ -1,4 +1,4 @@
-@import '../../web/styles/common';
+@import '../web/styles/common';
 
 /*
  * Input
