+++ src/web-search/SearchBox.module.css
@@ -1,4 +1,4 @@
-@import '../../web/styles/common';
+@import '../web/styles/common';
 
 /*
  * Search box
