+++ src/web-search/index.tsx
@@ -1,0 +1,2 @@
+// Main entry point
+export { default as LiveSearchInput } from './LiveSearchInput'
